LayoutTests/storage/indexeddb/setVersion-null-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB with null version string

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
db.setVersion(null);
PASS db.version is "null"
PASS successfullyParsed is true

TEST COMPLETE
